% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRSvIv.R
\name{SIRSvIv}
\alias{SIRSvIv}
\title{This function allows you to solve a simple vector borne diseases mathematical model
with delay (SIRSvIv). To try the solver please look at the example stated below.}
\usage{
SIRSvIv(pars = NULL, init = NULL, time = NULL, ...)
}
\arguments{
\item{tau}{=  Time delay}

\item{N}{= Total host population}

\item{beta}{= Host infection rate}

\item{r}{=  Host recovery rate}

\item{lambda}{= Number of new suceptible vectors per time}

\item{alpha}{= Vector infection rate}

\item{mu}{= Vector death rate}
}
\description{
Model
dS  = -beta Iv  S/N;
dI  = beta S(t-tau) Iv(t-tau) N - r I;
dR  = r I;
dSv = lambda - alpha I Sv/N - mu Sv;
dIv = alpha I Sv/N - mu Iv;
}
\examples{
parameters <- c(mu = 1 / (70 * 365), beta = 520 / 365,
                lambda = 1 / 14, alpha = 1 / 7, tau =1, N = 500, r = 1/6 )
initials <- c(S = 500,  I = 0, R = 0, Sv =0 , Iv = 1)
# Solve and plot.
S<- SIRSvIv(pars = parameters, init = initials, time = seq(from = 0, to = 10, by =.5))
plot(S$results)
}

